py/Boto3Project/requirements.txt
boto3==1.38.35
requests==2.32.4
PyYAML==6.0.2
